Move Sidebar component into sidebar file
