+++ src/common/components/index.ts
@@ -7,6 +7,7 @@ export { default as NavBar } from './generics/NavBar';
 export { default as Footer } from './generics/Footer';
 export { default as SelectLanguage } from './generics/SelectLanguage';
 export { default as ToggleThemeButton } from './generics/ToggleThemeButton';
+export { default as Sidebar } from './generics/sidebar/Sidebar';
 // Form Components
 export { default as Button } from './forms/Button';
 export { default as Input } from './forms/Input';
