+++ src/pages/admin/index.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Sidebar from '@src/common/components/generics/Sidebar';
+import Sidebar from '@src/common/components/generics/sidebar/Sidebar';
 
 export default function Admin() {
   return (
